refactor: shorten insight card file names
