+++ frontend/src/insights/InsightsPage.tsx
@@ -1,17 +1,17 @@
 import { useState } from 'react'
 import { useAuth } from '@/hooks/useAuth'
-import { CashFlowCard } from './components/cash-flow-card/CashFlowCard'
+import { CashFlowCard } from './components/cash-flow-card/Card'
 import {
   IncomeExpenseBreakdownCard,
   type BreakdownMode,
-} from './components/income-expense-breakdown-card/IncomeExpenseBreakdownCard'
+} from './components/income-expense-breakdown-card/Card'
 import { InsightsFloatingRangeControl } from './components/InsightsFloatingRangeControl'
-import { MerchantDistributionCard } from './components/merchant-distribution-card/MerchantDistributionCard'
-import { MerchantRankingCard } from './components/merchant-ranking-card/MerchantRankingCard'
-import { FundFlowCard } from './components/fund-flow-card/FundFlowCard'
-import { NetWorthCard } from './components/net-worth-card/NetWorthCard'
-import { PeriodGlanceCard } from './components/period-glance-card/PeriodGlanceCard'
-import { SavingsRateTrendCard } from './components/savings-rate-trend-card/SavingsRateTrendCard'
+import { MerchantDistributionCard } from './components/merchant-distribution-card/Card'
+import { MerchantRankingCard } from './components/merchant-ranking-card/Card'
+import { FundFlowCard } from './components/fund-flow-card/Card'
+import { NetWorthCard } from './components/net-worth-card/Card'
+import { PeriodGlanceCard } from './components/period-glance-card/Card'
+import { SavingsRateTrendCard } from './components/savings-rate-trend-card/Card'
 import { useInsightsCardData } from './hooks/useInsightsCardData'
 import { useInsightsCardQueries } from './hooks/useInsightsCardQueries'
 import { useInsightsCardVisibilityMap } from './hooks/useInsightsCardVisibilityMap'
