+++ frontend/src/insights/components/cash-flow-card/BarChart.tsx
