+++ frontend/src/insights/components/cash-flow-card/Card.tsx
@@ -6,7 +6,7 @@ import {
   LoadingOverlay,
 } from '@/components/LoadingTransition'
 import { useLoadingSnapshot } from '@/components/useLoadingSnapshot'
-import { CashFlowBarChart } from './CashFlowBarChart'
+import { CashFlowBarChart } from './BarChart'
 import type { CashFlowBarBucket, CashFlowGranularity } from '@/insights/types/cashFlow'
 import { getInsightsCashFlowFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
 import { formatSignedCurrency, getSignedAmountColor } from '@/insights/utils/money'
