+++ frontend/src/insights/components/fund-flow-card/Card.tsx
@@ -9,8 +9,8 @@ import type {
 import { getFundFlowChartHeight } from '@/insights/utils/fundFlowChart'
 import { withoutMatchingEntries } from '@/insights/utils/fundFlowEntries'
 import { getFundFlowFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
-import { FundFlowCategoryList } from './FundFlowCategoryList'
-import { FundFlowChart } from './FundFlowChart'
+import { FundFlowCategoryList } from './CategoryList'
+import { FundFlowChart } from './Chart'
 import { FxStatusBadge } from '../FxStatusBadge'
 import { InsightCalculationTooltip } from '../InsightCalculationTooltip'
 import { SectionHeader } from '../SectionHeader'
