+++ frontend/src/insights/components/fund-flow-card/CategoryList.tsx
