+++ frontend/src/insights/components/fund-flow-card/Chart.tsx
