+++ frontend/src/insights/components/income-expense-breakdown-card/TrendSections.tsx
