+++ frontend/src/insights/components/merchant-distribution-card/Legend.tsx
