+++ frontend/src/insights/components/merchant-distribution-card/MarketMap.tsx
