+++ frontend/src/insights/components/net-worth-card/Chart.tsx
