+++ frontend/src/insights/components/period-glance-card/display.ts
