+++ frontend/src/insights/components/savings-rate-trend-card/Chart.tsx
